Refatora autoconfiguracao para classes especificas, e implementa autoconfiguracao do JaxRsContractReader
