+++ java-restify-spring-autoconfigure/src/test/java/com/github/ljtfreitas/restify/spring/configure/EnableRestifyConfigurationTest.java
@@ -10,8 +10,8 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import com.github.ljtfreitas.restify.spring.configure.EnableRestify;
 import com.github.ljtfreitas.restify.spring.configure.EnableRestifyConfigurationTest.TestRestifyConfiguration;
+import com.github.ljtfreitas.restify.spring.whatever.TwitterApi;
 
 @RunWith(SpringRunner.class)
 @ContextConfiguration(classes = TestRestifyConfiguration.class)
@@ -26,7 +26,7 @@ public class EnableRestifyConfigurationTest {
 		assertNotNull(context.getBean(TwitterApi.class));
 	}
 
-	@EnableRestify
+	@EnableRestify(packages = "com.github.ljtfreitas.restify.spring.whatever")
 	static class TestRestifyConfiguration {
 	}
 }
