+++ java-restify-spring-autoconfigure/src/test/java/com/github/ljtfreitas/restify/spring/whatever/TwitterApi.java
@@ -1,4 +1,4 @@
-package com.github.ljtfreitas.restify.spring.configure;
+package com.github.ljtfreitas.restify.spring.whatever;
 
 import com.github.ljtfreitas.restify.spring.configure.Restifyable;
